build mode emits a link to the full trace to STDOUT (#5)
It would be nice for a URL to exist that will point directly to the trace representing this build.  This change adds output to `buildevents` when run in `build` mode (aka submitting the trace root span). The output from the command, if present, is a URL to the completed build. (If the API key fails to validate, the command is silent.)

```
➜  unset BUILDEVENT_APIKEY
➜  ./buildevents build myBuild 1558499469 success
#####  no valid API key, no output from builds
➜  export BUILDEVENT_APIKEY=$kiwi
➜  ./buildevents build myBuild 1558499469 success
https://ui.honeycomb.io/kiwi/datasets/buildevents/trace?trace_id=myBuild&trace_start_ts=1558499469&trace_end_ts=1558545868
#### ^^^ URL directly to the trace with end time 10 minutes in the future
```