add a command 'watch' that polls the circleci api (#14)
Given the difficulty in CircleCI of getting a job to always run (even when other jobs in the workflow fail), this change adds the functionality to poll the CircleCI API to get information about the status of the build. This new command `watch` is run in a job parallel to the rest of the build and it watches the build's status via API commands. When it detects that the build has finished, it sends along the root span along with the status of the build.

I forked the `go-circleci` SDK used to interact with the CircleCI API so that I could add some endpoints that exist only in v2 of the CirlceCI API - listing jobs in a workflow. https://github.com/jszwedko/go-circleci/pull/37 is open to merge those changes upstream, at which point `bulidevents` can stop using the fork.