Watch workflow status instead of jobs
This results in a much more consistent behaviour instead of querying for
jobs because CircleCI may take some time to start a new set of jobs
causing a race condition between the watch command and CircleCI starting
jobs. If we look at the workflow status then we know if the workflow
finished or not.