make watch job more efficient (#70)
keep track of whether or not we have blocked jobs.  this should let us know for certain when we're actually done, so we can stop waiting sooner.

We can stop and declare the build finished as soon as there are no running jobs (minus the watch job itself) and no blocked jobs.

Simplify the loop inside `evalWorkflow` to have no early returns, and just set another flag - `anyRunning`, which is true for jobs in either the `queued` or `running` circleci states.

The loop inside `waitCircle` continues to confound and is next on the list of places we should simplify, imo.