Add os and arch to the meta.* data sent
We live in a multiarch world now! I'm currently thinking about migrating
a bunch of CI runners over from x86->arm64, and I can see a few uses for
this in my buildevents data.

1. What parts of my CI runs are on an x86 vs arm executor? What's been
migrated, which spots are still to be done? (I expect this to be a
gradual migration as various dependencies - mostly docker images -
become available for arm.)

2. Performance improvements - can we compare run time of one arch versus
another?

Both of these will be easily queried in buildevents data if these fields
are present.

Since this PR uses runtime.{GOOS,GOARCH}, I believe performance impact
will be negligible - both of these are constants set by the golang
runtime on startup, we're just using them.