feat: return list of failed jobs in cmd watch (#225)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

This only applies to buildevents used in CircleCI.

In Honeycomb, we can query against the `watch` step of the buildevents
trace to find out how often jobs are failing. Unfortunately, this does
not include the list of the failed jobs, reducing its effectiveness, as
usually if you are looking at the amount of failed jobs you'd also like
to know which jobs failed.

As an example of reduced effectiveness, this prevents BubbleUp from
showing which jobs are responsible for a failing SLO on % succesful
builds.


![image](https://github.com/user-attachments/assets/809941c0-7a19-46a9-93e8-d0371eb8695f)

## Short description of the changes

We have the list of failed jobs available to us when we call the
circleCI API to find out if there are any failed jobs. This PR wires
through a list of names of those failed jobs back through the responses
so we can include it as a span attribute in the `watch` event.


## How to verify that this has the expected result

If we can test with this branch somehow, we should be able to see the
result in instrumentation. Let me know if you have ideas.